requirements.txt
streamlit==1.42.2
httpx==0.28.1
openai==1.64.0
streamlit-authenticator>=0.4.2
PyYAML==6.0.2